Add test case for firebase
